Move setup model function to separate modules
