+++ src/simcardems/cli.py
@@ -1,6 +1,4 @@
 import json
-import logging
-import os
 import typing
 from pathlib import Path
 
@@ -9,50 +7,16 @@ import click
 import dolfin
 from tqdm import tqdm
 
-from . import em_model
-from . import ep_model
 from . import mechanics_model
 from . import postprocess as post
 from . import save_load_functions as io
 from . import utils
 from .datacollector import DataCollector
-from .geometry import SlabGeometry
+from .setup_models import Defaults
+from .setup_models import setup_EM_model
 from .version import __version__
 
 logger = utils.getLogger(__name__)
-
-PathLike = typing.Union[os.PathLike, str]
-
-
-class _Defaults:
-    outdir: PathLike = "results"
-    T: float = 1000
-    dx: float = 0.2
-    dt: float = 0.05
-    bnd_cond: mechanics_model.BoundaryConditions = (
-        mechanics_model.BoundaryConditions.dirichlet
-    )
-    load_state: bool = False
-    cell_init_file: PathLike = ""
-    hpc: bool = False
-    lx: float = 2.0
-    ly: float = 0.7
-    lz: float = 0.3
-    save_freq: int = 1
-    pre_stretch: typing.Optional[typing.Union[dolfin.Constant, float]] = None
-    traction: typing.Union[dolfin.Constant, float] = None
-    spring: typing.Union[dolfin.Constant, float] = None
-    fix_right_plane: bool = True
-    loglevel = logging.INFO
-    num_refinements: int = 1
-    set_material: str = ""
-    drug_factors_file: str = ""
-    popu_factors_file: str = ""
-    disease_state: str = "healthy"
-
-
-def default_parameters():
-    return {k: v for k, v in _Defaults.__dict__.items() if not k.startswith("_")}
 
 
 class _tqdm:
@@ -76,73 +40,73 @@ def cli():
 @click.option(
     "-o",
     "--outdir",
-    default=_Defaults.outdir,
+    default=Defaults.outdir,
     type=click.Path(writable=True, resolve_path=True),
     help="Output directory",
 )
-@click.option("--dt", default=_Defaults.dt, type=float, help="Time step")
+@click.option("--dt", default=Defaults.dt, type=float, help="Time step")
 @click.option(
     "-T",
     "--end-time",
     "T",
-    default=_Defaults.T,
+    default=Defaults.T,
     type=float,
     help="End-time of simulation",
 )
 @click.option(
     "-n",
     "--num_refinements",
-    default=_Defaults.num_refinements,
+    default=Defaults.num_refinements,
     type=int,
     help="Number of refinements of for the mesh using in the EP model",
 )
 @click.option(
     "--save_freq",
-    default=_Defaults.save_freq,
+    default=Defaults.save_freq,
     type=int,
     help="Set frequency of saving results to file",
 )
 @click.option(
     "--set_material",
-    default=_Defaults.set_material,
+    default=Defaults.set_material,
     type=str,
     help="Choose material properties for mechanics model (default is HolzapfelOgden, option is Guccione",
 )
-@click.option("-dx", default=_Defaults.dx, type=float, help="Spatial discretization")
+@click.option("-dx", default=Defaults.dx, type=float, help="Spatial discretization")
 @click.option(
     "-lx",
-    default=_Defaults.lx,
+    default=Defaults.lx,
     type=float,
     help="Size of mesh in x-direction",
 )
 @click.option(
     "-ly",
-    default=_Defaults.ly,
+    default=Defaults.ly,
     type=float,
     help="Size of mesh in y-direction",
 )
 @click.option(
     "-lz",
-    default=_Defaults.lz,
+    default=Defaults.lz,
     type=float,
     help="Size of mesh in z-direction",
 )
 @click.option(
     "--bnd_cond",
-    default=_Defaults.bnd_cond,
+    default=Defaults.bnd_cond,
     type=click.Choice(mechanics_model.BoundaryConditions._member_names_),
     help="Boundary conditions for the mechanics problem",
 )
 @click.option(
     "--load_state",
     is_flag=True,
-    default=_Defaults.load_state,
+    default=Defaults.load_state,
     help="If load existing state if exists, otherwise create a new state",
 )
 @click.option(
     "-IC",
     "--cell_init_file",
-    default=_Defaults.cell_init_file,
+    default=Defaults.cell_init_file,
     type=str,
     help=(
         "Path to file containing initial conditions (json or h5 file). "
@@ -151,42 +115,42 @@ def cli():
 )
 @click.option(
     "--loglevel",
-    default=_Defaults.loglevel,
+    default=Defaults.loglevel,
     type=int,
     help="How much printing. DEBUG: 10, INFO:20 (default), WARNING: 30",
 )
 @click.option(
     "--hpc",
     is_flag=True,
-    default=_Defaults.hpc,
+    default=Defaults.hpc,
     help="Indicate if simulations runs on hpc. This turns off the progress bar.",
 )
 @click.option(
     "--drug_factors_file",
-    default=_Defaults.drug_factors_file,
+    default=Defaults.drug_factors_file,
     type=str,
     help="Set drugs scaling factors (json file)",
 )
 @click.option(
     "--popu_factors_file",
-    default=_Defaults.popu_factors_file,
+    default=Defaults.popu_factors_file,
     type=str,
     help="Set population scaling factors (json file)",
 )
 @click.option(
     "--disease_state",
-    default=_Defaults.disease_state,
+    default=Defaults.disease_state,
     type=str,
     help="Indicate disease state. Default is healthy. ",
 )
 def run(
-    outdir: PathLike,
+    outdir: utils.PathLike,
     T: float,
     dx: float,
     dt: float,
     bnd_cond: mechanics_model.BoundaryConditions,
     load_state: bool,
-    cell_init_file: PathLike,
+    cell_init_file: utils.PathLike,
     hpc: bool,
     lx: float,
     ly: float,
@@ -231,28 +195,28 @@ def run_json(path):
 
 
 def main(
-    outdir: PathLike = _Defaults.outdir,
-    T: float = _Defaults.T,
-    dx: float = _Defaults.dx,
-    dt: float = _Defaults.dt,
-    bnd_cond: mechanics_model.BoundaryConditions = _Defaults.bnd_cond,
-    load_state: bool = _Defaults.load_state,
-    cell_init_file: PathLike = _Defaults.cell_init_file,
-    hpc: bool = _Defaults.hpc,
-    lx: float = _Defaults.lx,
-    ly: float = _Defaults.ly,
-    lz: float = _Defaults.lz,
-    save_freq: int = _Defaults.save_freq,
+    outdir: utils.PathLike = Defaults.outdir,
+    T: float = Defaults.T,
+    dx: float = Defaults.dx,
+    dt: float = Defaults.dt,
+    cell_init_file: utils.PathLike = Defaults.cell_init_file,
+    lx: float = Defaults.lx,
+    ly: float = Defaults.ly,
+    lz: float = Defaults.lz,
     pre_stretch: typing.Optional[typing.Union[dolfin.Constant, float]] = None,
     traction: typing.Union[dolfin.Constant, float] = None,
     spring: typing.Union[dolfin.Constant, float] = None,
     fix_right_plane: bool = True,
-    loglevel: int = _Defaults.loglevel,
-    num_refinements: int = _Defaults.num_refinements,
-    set_material: str = _Defaults.set_material,
-    drug_factors_file: str = _Defaults.drug_factors_file,
-    popu_factors_file: str = _Defaults.popu_factors_file,
-    disease_state: str = _Defaults.disease_state,
+    num_refinements: int = Defaults.num_refinements,
+    set_material: str = Defaults.set_material,
+    bnd_cond: mechanics_model.BoundaryConditions = Defaults.bnd_cond,
+    load_state: bool = Defaults.load_state,
+    hpc: bool = Defaults.hpc,
+    save_freq: int = Defaults.save_freq,
+    loglevel: int = Defaults.loglevel,
+    drug_factors_file: str = Defaults.drug_factors_file,
+    popu_factors_file: str = Defaults.popu_factors_file,
+    disease_state: str = Defaults.disease_state,
 ):
 
     # Get all arguments and dump them to a json file
@@ -283,41 +247,20 @@ def main(
     else:
         logger.info("Create a new state")
         # Create a new state
-        geometry = SlabGeometry(
+        coupling, solver, mech_heart, t0 = setup_EM_model(
+            dx=dx,
+            dt=dt,
+            bnd_cond=bnd_cond,
+            cell_init_file=cell_init_file,
             lx=lx,
             ly=ly,
             lz=lz,
-            dx=dx,
             num_refinements=num_refinements,
-        )
-
-        coupling = em_model.EMCoupling(geometry)
-
-        # Set-up solver and time it
-        solver = ep_model.setup_solver(
-            dt=dt,
-            coupling=coupling,
-            cell_init_file=cell_init_file,
+            set_material=set_material,
             drug_factors_file=drug_factors_file,
             popu_factors_file=popu_factors_file,
             disease_state=disease_state,
         )
-
-        coupling.register_ep_model(solver)
-
-        with dolfin.Timer("[demo] Setup Mech solver"):
-            mech_heart = mechanics_model.setup_mechanics_model(
-                coupling=coupling,
-                dt=dt,
-                bnd_cond=bnd_cond,
-                cell_params=solver.ode_solver._model.parameters(),
-                pre_stretch=pre_stretch,
-                traction=traction,
-                spring=spring,
-                fix_right_plane=fix_right_plane,
-                set_material=set_material,
-            )
-        t0 = 0
 
     logger.info(f"Starting at t0={t0}")
