+++ src/simcardems/ep_model.py
@@ -195,56 +195,3 @@ def handle_cell_inits(
     if cell_inits is not None:
         cell_inits_tmp.update(cell_inits)
     return cell_inits_tmp
-
-
-def setup_solver(
-    dt,
-    coupling,
-    scheme="GRL1",
-    theta=0.5,
-    preconditioner="sor",
-    cell_params=None,
-    cell_inits=None,
-    cell_init_file=None,
-    drug_factors_file=None,
-    popu_factors_file=None,
-    disease_state="healthy",
-):
-    ps = setup_splitting_solver_parameters(
-        theta=theta,
-        preconditioner=preconditioner,
-        dt=dt,
-        scheme=scheme,
-    )
-
-    cell_params = handle_cell_params(
-        cell_params=cell_params,
-        disease_state=disease_state,
-        drug_factors_file=drug_factors_file,
-        popu_factors_file=popu_factors_file,
-    )
-
-    cell_inits = handle_cell_inits(cell_inits=cell_inits, cell_init_file=cell_init_file)
-
-    cell_inits["lmbda"] = coupling.lmbda_ep
-    cell_inits["Zetas"] = coupling.Zetas_ep
-    cell_inits["Zetaw"] = coupling.Zetaw_ep
-
-    cellmodel = CellModel(init_conditions=cell_inits, params=cell_params)
-
-    # Set-up cardiac model
-    ep_heart = setup_ep_model(cellmodel, coupling.ep_mesh)
-    timer = dolfin.Timer("SplittingSolver: setup")
-
-    solver = cbcbeat.SplittingSolver(ep_heart, ps)
-
-    timer.stop()
-    # Extract the solution fields and set the initial conditions
-    (vs_, vs, vur) = solver.solution_fields()
-    vs_.assign(cellmodel.initial_conditions())
-
-    # Output some degrees of freedom
-    total_dofs = vs.function_space().dim()
-    logger.info("EP model")
-    utils.print_mesh_info(coupling.ep_mesh, total_dofs)
-    return solver
